삭제 Confirm 추가 Delete confirmation dialog in Blazor https://www.youtube.com/watch?v=Caw5hmq4dEY&list=PL6n9fhu94yhVowClAs8-6nYnfsOTma14P&index=45

+++ EmployeeManagement.Web/Pages/EditEmployeeBase.cs
@@ -93,11 +93,27 @@ namespace EmployeeManagement.Web.Pages
             }
         }
 
-        protected async Task Delete_Click()
+        protected ProgimTech.Components.ConfirmBase DeleteConfirmation { get; set; }
+
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            NavigationManager.NavigateTo("/");
+            DeleteConfirmation.Show();
         }
+
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                NavigationManager.NavigateTo("/");
+            }
+        }
+
+        //protected async Task Delete_Click()
+        //{
+        //    await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+        //    NavigationManager.NavigateTo("/");
+        //}
 
     }
 }
